added getStatistics feature

+++ server/src/services/bank.service.ts
@@ -1,13 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable, forwardRef } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Bank } from '../models/bank.model';
+import { TransactionService } from './transaction.service';
 
 @Injectable()
 export class BankService {
   constructor(
     @InjectRepository(Bank)
     private bankRepository: Repository<Bank>,
+    @Inject(forwardRef(() => TransactionService))
+    private transactionService: TransactionService,
   ) {}
 
   findAll(): Promise<Bank[]> {
@@ -22,8 +25,13 @@ export class BankService {
     return await this.bankRepository.save(this.bankRepository.create(args));
   }
 
-  async delete(id: number): Promise<void> {
+  async delete(id: number): Promise<Bank> {
+    const bank = await this.bankRepository.findOneBy({ id });
+    const transaction = await this.transactionService.findOneByBankId(id);
+    if (transaction)
+      throw new Error('You can`t delete bank if it has any transactions');
     await this.bankRepository.delete(id);
+    return bank;
   }
 
   async update(args: Bank): Promise<Bank> {
