+++ server/src/services/category.service.ts
@@ -1,13 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable, forwardRef } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { In, Repository } from 'typeorm';
 import { Category } from '../models/category.model';
+import { TransactionService } from './transaction.service';
 
 @Injectable()
 export class CategoryService {
   constructor(
     @InjectRepository(Category)
     private categoryRepository: Repository<Category>,
+    @Inject(forwardRef(() => TransactionService))
+    private transactionService: TransactionService,
   ) {}
 
   findAll(): Promise<Category[]> {
@@ -26,8 +29,15 @@ export class CategoryService {
     return this.categoryRepository.save(this.categoryRepository.create(args));
   }
 
-  async delete(id: number): Promise<void> {
+  async delete(id: number): Promise<Category> {
+    const category = await this.categoryRepository.findOneBy({ id });
+    const transaction = await this.transactionService.findOneByCategoryId(id);
+    if (transaction)
+      throw new Error(
+        'You can`t delete category if it has relations with any transactions',
+      );
     await this.categoryRepository.delete(id);
+    return category;
   }
 
   async update(args: Category): Promise<Category> {
